+++ R/xtable_to_flextable.R
@@ -114,9 +114,9 @@ as_flextable.xtable <- function(
     format.args$decimal.mark <- options()$OutDec
   }
 
-  digits_val <- attr(x, "digits", exact = TRUE)
-  display_val <- attr(x, "display", exact = TRUE)
-  align <- attr(x, "align")
+  digits_val <- attr2(x, "digits")
+  display_val <- attr2(x, "display")
+  align <- attr2(x, "align")
 
   if (!include.rownames) {
     digits_val <- digits_val[-1]